+++ internal/provider/resource_ise_network_access_authentication_rule.go
@@ -36,6 +36,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/netascode/go-ise"
+	"github.com/tidwall/gjson"
 )
 
 //template:end imports
@@ -266,7 +267,6 @@ func (r *NetworkAccessAuthenticationRuleResource) Configure(_ context.Context, r
 
 //template:end model
 
-//template:begin create
 func (r *NetworkAccessAuthenticationRuleResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	var plan NetworkAccessAuthenticationRule
 
@@ -281,20 +281,40 @@ func (r *NetworkAccessAuthenticationRuleResource) Create(ctx context.Context, re
 
 	// Create object
 	body := plan.toBody(ctx, NetworkAccessAuthenticationRule{})
-	res, _, err := r.client.Post(plan.getPath(), body)
-	if err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (POST), got error: %s, %s", err, res.String()))
-		return
+	if plan.Name.ValueString() != "Default" {
+		res, _, err := r.client.Post(plan.getPath(), body)
+		if err != nil {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (POST), got error: %s, %s", err, res.String()))
+			return
+		}
+		plan.Id = types.StringValue(res.Get("response.rule.id").String())
+	} else {
+		res, err := r.client.Get(plan.getPath())
+		if err != nil {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve objects, got error: %s", err))
+			return
+		}
+		if value := res.Get("response"); len(value.Array()) > 0 {
+			value.ForEach(func(k, v gjson.Result) bool {
+				if v.Get("rule.name").String() == plan.Name.ValueString() {
+					plan.Id = types.StringValue(v.Get("rule.id").String())
+					return false
+				}
+				return true
+			})
+		}
+		res, err = r.client.Put(plan.getPath()+"/"+plan.Id.ValueString(), body)
+		if err != nil {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (PUT), got error: %s, %s", err, res.String()))
+			return
+		}
 	}
-	plan.Id = types.StringValue(res.Get("response.rule.id").String())
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Create finished successfully", plan.Id.ValueString()))
 
 	diags = resp.State.Set(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 }
-
-//template:end create
 
 //template:begin read
 func (r *NetworkAccessAuthenticationRuleResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {